+++ app/config/services_test.yml
@@ -1,5 +1,5 @@
 services:
   app_test.swagger_spec_validator:
-    class: AppBundle\Tests\Controller\SwaggerSpecValidator
+    class: AppBundle\Tests\Controller\SwaggerValidator\SwaggerSpecValidator
     arguments:
       - "%kernel.root_dir%/../doc/theatre.json"
