+++ src/AppBundle/Tests/Controller/AbstractController.php
@@ -81,7 +81,7 @@ abstract class AbstractController extends WebTestCase
 
         $uri = parse_url($this->getClient()->getRequest()->getRequestUri(), PHP_URL_PATH);
 
-        if ((substr($uri, 0,7) != '/admin/') && (substr($uri, 0,5) != '/doc/') && (substr($uri, 0,6) != '/tags/')) {
+        if ((substr($uri, 0, 7) != '/admin/') && (substr($uri, 0, 5) != '/doc/') && (substr($uri, 0, 6) != '/tags/')) {
             $this->getContainer()->get('app_test.swagger_spec_validator')->assertResource(
                 $this->getContainer()->get('router')->match($uri)['_route'],
                 $this->getClient()->getRequest(),
