+++ src/AppBundle/Tests/Controller/SwaggerValidator/Parameters/AbstractParameterValidator.php
@@ -2,7 +2,6 @@
 
 namespace AppBundle\Tests\Controller\SwaggerValidator\Parameters;
 
-
 abstract class AbstractParameterValidator extends \PHPUnit_Framework_Assert
 {
     /**
