+++ src/AppBundle/Tests/Controller/SwaggerValidator/Parameters/PathParameter/IntegerPathParameterValidator.php
@@ -9,7 +9,7 @@ class IntegerPathParameterValidator extends AbstractParameterValidator
 {
     /**
      * @param AbstractParameter $parameterDoc
-     * @param integer $parameterRequest
+     * @param $parameterRequest
      */
     public function validate(AbstractParameter $parameterDoc, $parameterRequest)
     {
@@ -21,8 +21,8 @@ class IntegerPathParameterValidator extends AbstractParameterValidator
         if ($parameterRequest === null) {
             self::assertFalse($parameterDoc->isRequired());
             self::assertNotFalse($parameterDoc->getDefault());
-        } else {
-            self::assertRegExp('/^\d+$/', $parameterRequest);
+            $parameterRequest = $parameterDoc->getDefault();
         }
+        self::assertRegExp('/^\d+$/', $parameterRequest);
     }
 }
